Decoupling Order Repository from Order Service
"D" in SOLID states that high level modules should not depend on low
 level modules, they should depend on abstractions. Order Service
should depend on the abstraction of Order Repository.

Order Service also should not be responsible for initialization of Order
Repository, we should inject the dependency in the constructor,
ideally using IoC conatiner.

We will do the same about Customer Repository in the next commit.