Extract Order Validation Rules from Service
"S" in SOLID states that classes should have one and only one
reason to change. Order Service should only hold details about
different steps that need to be taken in order to place an Order and
Order validation details are to specific for this level. This will protect Service
class from the possible future changes in Order validation.