Reorganize project files
Abstractions were placed into the API folder so that later if we will
need different implementations we can easily move them into a
separate assembly that will be referenced by multiple projects.

All elements with public access modifiers were kept in the original
namespace so that we do not break any existing usages of this libraries.