Place an order, linked to a customer
An association is made between Order and Customer Object so that
Order/Customer relationship can be persisted to the database.
An alternative solution could be to pass the Customer to the Save
method on the Order Repository, but this woul mean we need to
change a method signature on our Repository abstraction which
we want to avoid.

As mentioned in the code comment ideally we do not want the
Order/Customer association to be done in Service class. Maybe we
can refactor it later and put into the Order Repository class.