Load all customers and their linked orders
Plural nouns were used as table names (Customers, Orders).

Orders association was added to the Customer entity. Since we can only add
new orders through the Save method on the repository, a read only collection
was used.

A reference to the Customer primary key was added to Order so that it is easier
to build an object graph of Customers and their Orders. Not ideal solution since
this is a database concept.

A new method has been added for loading and associating all the Orders with all the
Customers. Needs attention (not scalable!) because we are retrieving everything from
the database and then working in memory. Dapper multiple results approach was used
an alternative would be Multi Mapping with a sql join, the downsides are data is not normalised,
advantage is there would be less iterations to build a graph after sql retrieval.