Initialization of Order Service is moved to a factory
This service can't be directly initialized because it depends on an
internal component (Order Repository). In order to respect the access
modifier a public factory class was created which will enable callers from
outside this assembly to initialize the service.

It would also be useful to discuss why the internal modifier on the repository.