Create a demo project
The project shows how to wire-up all the components to:
- Load all customers and their linked orders
- Place an order, linked to a customer

Further work will have to be done on the PlaceOrder method because it is not intuitive at the moment. Caller needs to know its inner details to use it successfully and lacks of information about what went wrong if the method returns false.

No existing method signatures or access modifiers were changed during this refactoring session.

For automatic resolution of components an IoC container can be use (idea for future refactoring).